Add ffi_sample.cpp to demonstrate calling C# functions from a DLL and update PaymentFoundation class for correct method signatures

+++ src/MethodOverloading/FinanceLib/PaymentFoundation.cs
@@ -15,15 +15,15 @@ public class PaymentFoundation
     }
 
     // Publish error. The method signatures are the same, so we need to give them different entry points.
-    
+
     // [UnmanagedCallersOnly(EntryPoint = "ProcessPayment")]
-    // public static void ExportProcessPayment(decimal amount) => ProcessPayment(amount);
+    // public static void ExportProcessPayment(double amount) => ProcessPayment((decimal)amount);
 
     // [UnmanagedCallersOnly(EntryPoint = "ProcessPayment")]
     // public static void ExportProcessPayment(int bonus) => ProcessPayment(bonus);
 
-        [UnmanagedCallersOnly(EntryPoint = "ProcessPayment_WithAmount")]
-    public static void ExportProcessPayment(decimal amount) => ProcessPayment(amount);
+    [UnmanagedCallersOnly(EntryPoint = "ProcessPayment_WithAmount")]
+    public static void ExportProcessPayment(double amount) => ProcessPayment((decimal)amount);
 
     [UnmanagedCallersOnly(EntryPoint = "ProcessPayment_WithBonus")]
     public static void ExportProcessPayment(int bonus) => ProcessPayment(bonus);
